% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinput.R
\name{checkinput}
\alias{checkinput}
\title{Check the user input for issues}
\usage{
checkinput(tree, nspec, samplesizes, popsizeprior, divtimeprior)
}
\arguments{
\item{tree}{A guide tree specified in Newick format. Tip labels should be 0 to (nspec-1).}

\item{nspec}{The number of potential species.}

\item{samplesizes}{A vector of sample sizes.}

\item{popsizeprior}{Prior for population sizes.}

\item{divtimeprior}{Prior on divergence times.}
}
\description{
Check the user input for issues
}
