% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaner.R
\name{clean_working}
\alias{clean_working}
\title{Remove unnecessary files from the working directory}
\usage{
clean_working(prefix)
}
\arguments{
\item{prefix}{The prefix used in the delimitR analysis you wish to delete files from.}
}
\description{
Remove unnecessary files from the working directory
}
