% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{fastsimcoalsims}
\alias{fastsimcoalsims}
\title{Simulate data in fastsimcoal2}
\usage{
fastsimcoalsims(prefix, pathtofsc, nreps)
}
\arguments{
\item{prefix}{The prefix for the .tpl and .est files in your working directory.}

\item{pathtofsc}{The path to fastsimcoal2.}

\item{nreps}{The number of datasets to simulate under each model.}
}
\value{
This will simulate data in directories created in your working directory.
}
\description{
Simulate data in fastsimcoal2
}
