% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhistory_v2.R
\name{generatemigratelist}
\alias{generatemigratelist}
\title{Generate history}
\usage{
generatemigratelist(history, parsedtree, migrationmatrix)
}
\description{
Generate history
}
\keyword{internal}
