% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhistory.R
\name{generatemigratelist_old}
\alias{generatemigratelist_old}
\title{Generate history}
\usage{
generatemigratelist_old(history, parsedtree, migrationmatrix)
}
\description{
Generate history
}
\keyword{internal}
