% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createprior.R
\name{makeprior}
\alias{makeprior}
\title{Build a prior}
\usage{
makeprior(prefix, nspec, nclasses, mydir, traitsfile, threshold, thefolder,
  ncores)
}
\arguments{
\item{prefix}{The prefix for the .tpl and .est files in your working directory.}

\item{nspec}{The number of putative species.}

\item{nclasses}{The number of bins to use in the SFS.}

\item{mydir}{The working directory for the analysis.}

\item{traitsfile}{A beast-formatted traits file.}

\item{threshold}{If a downsampling approach was used to create the observed SFS, provide the threshold here.}

\item{thefolder}{A name for the folder to save the prior in.}

\item{ncores}{The number of cores available for the analysis.}
}
\value{
This will return the full prior to use when building the RF classifier.
}
\description{
Build a prior
}
