% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeparser.R
\name{parsetree}
\alias{parsetree}
\title{Generate a list of lists that containes information about coalescent events from a guide tree.}
\usage{
parsetree(tree)
}
\arguments{
\item{tree}{A guide tree.}
}
\description{
Generate a list of lists that containes information about coalescent events from a guide tree.
}
