% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepobs.R
\name{prepobserved}
\alias{prepobserved}
\title{Prepare the observed SFS}
\usage{
prepobserved(observed, FullPrior, ReducedPrior, nclasses, npops,
  traitsfile, threshold)
}
\arguments{
\item{observed}{This should be an observed multi-dimensional Site Frequency Spectrum as implemented in fastsimcoal2. Please ensure that you follow instructions from the package vignette concerning population naming and order.}

\item{FullPrior}{This should be the full prior as generated by the function \code{\link{makeprior}}.}

\item{ReducedPrior}{This should be the prior with invariant columns removed, as created by the function \code{\link{Prior_reduced}} \link{Prior_reduced}.}

\item{nclasses}{This is the number of classes that should be used to construct the binned mSFS, and should match that used for the Prior.}

\item{npops}{This should be equal to the number of potential species included in the analysis.}

\item{traitsfile}{This should be the name of the traits file for the observed data. Formatting follows the vignette.}

\item{threshold}{The percentage of individuals used to build the mSFS. This is relevant if you used a downsampling approach to build the mSFS from the observed data, but used the original number of individuals for generating models in delimitR.}
}
\description{
Prepare the observed SFS
}
