% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFbuild_abcrf.R
\name{Prior_reduced}
\alias{Prior_reduced}
\title{Prepare the reduced SFS by removing bins of the SFS with no variance.s}
\usage{
Prior_reduced(FullPrior)
}
\arguments{
\item{FullPrior}{This should be the full prior as generated by the function \code{\link{makeprior}} \link{makeprior}.}
}
\description{
Prepare the reduced SFS by removing bins of the SFS with no variance.s
}
