% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFbuild_abcrf.R
\name{RF_build_abcrf}
\alias{RF_build_abcrf}
\title{Build the RF classifier.}
\usage{
RF_build_abcrf(ReducedPrior, FullPrior, ntrees)
}
\arguments{
\item{ReducedPrior}{This should be the prior with invariant columns removed, as created by the function \code{\link{Prior_reduced}} \link{Prior_reduced}.}

\item{FullPrior}{This should be the full prior as generated by the function \code{\link{makeprior}} \link{makeprior}.}

\item{ntrees}{The number of trees used to build the RF classifier.}
}
\description{
Build the RF classifier.
}
