% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFbuild_abcrf.R
\name{RF_predict_abcrf}
\alias{RF_predict_abcrf}
\title{Predict the model for the observed data}
\usage{
RF_predict_abcrf(RF, Observed_Reduced, ReducedPrior, FullPrior, ntrees)
}
\arguments{
\item{RF}{The RF classifier built using the function \code{\link{RF_build_abcrf}} \link{RF_build_abcrf}.}

\item{Observed_Reduced}{The observed data as prepared using the function \code{\link{prepobserved}} \link{prepobserved}.}

\item{ReducedPrior}{This should be the prior with invariant columns removed, as created by the function \link{Prior_reduced}.}

\item{FullPrior}{This should be the full prior as generated by the function \code{\link{makeprior}} \link{makeprior}.}

\item{ntrees}{The number of trees to use in the RF classifier.}
}
\description{
Predict the model for the observed data
}
