% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterfunction.R
\name{setup_fsc2}
\alias{setup_fsc2}
\title{Generate input files for fastsimcoal2}
\usage{
setup_fsc2(tree, samplesizes, nspec, popsizeprior, divtimeprior,
  migrateprior, migmatrix, nsnps, prefix, secondarycontact, divwgeneflow,
  maxmigrations, myrules = NULL)
}
\arguments{
\item{tree}{A guide tree specified in Newick format. Tip labels should be 0 to (nspec-1).}

\item{samplesizes}{A vector of sample sizes.}

\item{nspec}{The number of potential species.}

\item{popsizeprior}{Prior for population sizes.}

\item{divtimeprior}{Prior on divergence times.}

\item{migrateprior}{Prior on migration rates.}

\item{migmatrix}{Matrix specifying which species migration can happen between.}

\item{nsnps}{The number of SNPs to simulate.}

\item{prefix}{A prefix for naming output filess.}

\item{secondarycontact}{Boolean indicating whether or not to include secondary contact in models.}

\item{divwgeneflow}{Boolean indicating whether or not to include divergence with gene flow in models.}

\item{maxmigrations}{The maximum number of migration events to include in any single model.}

\item{myrules}{A list of rules for the order of coalescence times.}
}
\description{
Generate input files for fastsimcoal2
}
