% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeest.R
\name{writeest}
\alias{writeest}
\title{Generate .est files}
\usage{
writeest(samplesizes, nhistevents, nsnps, prefix, myhistoricinfo,
  popsizeprior, divtimeprior, migrateprior, myrules)
}
\description{
Generate .est files
}
\keyword{internal}
